src/commands/general.py
from discord.ext import commands

class GeneralCommands(commands.Cog):
    def __init__(self, bot):
        self.bot = bot

    @commands.command()
    async def hello(self, ctx):
        await ctx.send("Hello from discord!")

def setup(bot):
    bot.add_cog(GeneralCommands(bot))
